NEW - # 2276: Intersection of circle and (function) line not on circle https://dev.geogebra.org/trac/ticket/2276
Suppose f(x) = b (x-1) + 0.5 = b x + (0.5 - b), and b=1 at the moment,

when AlgoIntersectPolynomial tries to use this function, Function.expandToPolyFunctionNoCas(ExpressionValue ev, boolean symbolic) will attempt to change all the NumberValue to its double value, hence b will be successfully changed to "1". However, "(0.5-b)" is an ExpressionNode but not a NumberValue, hence it is changed to "0"  (although -0.5 is the correct value of it)

So I added a condition: if the a coefficient is an ExpressionNode, then call evaluateNum().getDouble() to get the correct value.

git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@17922 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e