ASSIGNED - # 1044: Intersection point: Special Case Handling https://dev.geogebra.org/trac/ticket/1044
Removed Probabilistic Checking; checked many of the past tickets and it works for all of them. Need some more testing on various cases, and also need to deal with the incident points of e.g. intersect line and polygon (i.e. add the intersection points of line and polygon to the incidence list of all the intersection points)

git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@17990 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e